refactor(vitrine): split wasm into cohesive modules
