+++ crates/vize_vitrine/src/wasm/compiler.rs
@@ -1,36 +1,6 @@
-//! WASM bindings for Vue compiler.
-//!
-//! FFI boundary code: uses std types for JavaScript interop.
-#![allow(
-    clippy::disallowed_types,
-    clippy::disallowed_methods,
-    clippy::disallowed_macros
-)]
+//! The `Compiler` WASM class, its free-function aliases, and the internal
+//! template/SFC compilation pipeline.
 
-mod analyze;
-mod cross_file;
-#[cfg(feature = "glyph")]
-mod format;
-mod inspector;
-mod lint;
-mod musea;
-
-// Re-export type checking bindings from separate module
-#[path = "wasm_typecheck.rs"]
-mod wasm_typecheck;
-
-// Re-export all WASM bindings
-pub use analyze::*;
-pub use cross_file::*;
-#[cfg(feature = "glyph")]
-pub use format::*;
-pub use inspector::*;
-pub use lint::*;
-pub use musea::*;
-pub use wasm_typecheck::*;
-
-use serde::Serialize;
-use std::collections::BTreeMap;
 use vize_carton::Bump;
 use wasm_bindgen::prelude::*;
 
@@ -40,405 +10,18 @@ use vize_atelier_core::parser::parse;
 use vize_atelier_dom::{DomCompilerOptions, compile_template_with_template_syntax};
 use vize_atelier_sfc::compile_script::typescript::transform_typescript_to_js;
 use vize_atelier_sfc::{
-    CssCompileOptions, CssTargets, ScriptCompileOptions, SfcCompileOptions, SfcDescriptor,
-    SfcMacroArtifact, SfcParseOptions, StyleCompileOptions, TemplateCompileOptions,
-    compile_sfc_with_template_syntax as sfc_compile_with_template_syntax, parse_sfc,
+    ScriptCompileOptions, SfcCompileOptions, SfcParseOptions, StyleCompileOptions,
+    TemplateCompileOptions, compile_sfc_with_template_syntax as sfc_compile_with_template_syntax,
+    parse_sfc,
 };
 use vize_atelier_ssr::{SsrCompilerOptions, compile_ssr_with_template_syntax};
 use vize_atelier_vapor::{VaporCompilerOptions, compile_vapor_with_template_syntax};
 
-/// Helper function to serialize values to JsValue with maps as objects
-pub(crate) fn to_js_value<T: Serialize>(value: &T) -> Result<JsValue, JsValue> {
-    let serializer = serde_wasm_bindgen::Serializer::new().serialize_maps_as_objects(true);
-    value
-        .serialize(&serializer)
-        .map_err(|e| JsValue::from_str(&e.to_string()))
-}
-
-fn to_json_js_value<T: Serialize>(value: &T) -> Result<JsValue, JsValue> {
-    let json = serde_json::to_string(value).map_err(|e| JsValue::from_str(&e.to_string()))?;
-    js_sys::JSON::parse(&json)
-}
-
-struct ParsedCompilerOptions {
-    options: CompilerOptions,
-    binding_metadata: Option<vize_atelier_core::options::BindingMetadata>,
-}
-
-fn parse_compiler_options(options: &JsValue) -> ParsedCompilerOptions {
-    let get_string = |key: &str| {
-        js_sys::Reflect::get(options, &JsValue::from_str(key))
-            .ok()
-            .and_then(|value| value.as_string())
-    };
-
-    let get_bool = |key: &str| {
-        js_sys::Reflect::get(options, &JsValue::from_str(key))
-            .ok()
-            .and_then(|value| value.as_bool())
-    };
-
-    let binding_metadata = js_sys::Reflect::get(options, &JsValue::from_str("bindingMetadata"))
-        .ok()
-        .and_then(|value| {
-            if value.is_null() || value.is_undefined() {
-                return None;
-            }
-            let json = js_sys::JSON::stringify(&value).ok()?.as_string()?;
-            serde_json::from_str(&json).ok()
-        });
-
-    ParsedCompilerOptions {
-        options: CompilerOptions {
-            mode: get_string("mode"),
-            prefix_identifiers: get_bool("prefixIdentifiers"),
-            hoist_static: get_bool("hoistStatic"),
-            cache_handlers: get_bool("cacheHandlers"),
-            scope_id: get_string("scopeId"),
-            ssr: get_bool("ssr"),
-            source_map: get_bool("sourceMap"),
-            filename: get_string("filename"),
-            output_mode: get_string("outputMode"),
-            is_ts: get_bool("isTs"),
-            custom_renderer: get_bool("customRenderer"),
-            template_syntax: get_string("templateSyntax"),
-            runtime_module_name: get_string("runtimeModuleName"),
-            runtime_global_name: get_string("runtimeGlobalName"),
-            script_ext: get_string("scriptExt"),
-        },
-        binding_metadata,
-    }
-}
-
-/// Convert a UTF-8 byte offset to a JavaScript string offset (UTF-16 code units).
-/// OXC and the SFC parser report byte offsets, while Monaco/JS consumers expect
-/// UTF-16 offsets.
-pub(crate) fn utf8_byte_to_utf16_offset(content: &str, byte_offset: u32) -> u32 {
-    let byte_offset = byte_offset as usize;
-    if byte_offset >= content.len() {
-        return content.encode_utf16().count() as u32;
-    }
-    content[..byte_offset].encode_utf16().count() as u32
-}
-
-#[cfg(test)]
-mod tests {
-    use super::utf8_byte_to_utf16_offset;
-
-    #[test]
-    fn test_utf8_byte_to_utf16_offset_handles_multibyte_and_astral_chars() {
-        let source = "aあ😀b";
-
-        let hiragana_start = source.find('あ').expect("hiragana should exist") as u32;
-        let emoji_start = source.find('😀').expect("emoji should exist") as u32;
-        let latin_b_start = source.find('b').expect("latin b should exist") as u32;
-
-        assert_eq!(utf8_byte_to_utf16_offset(source, 0), 0);
-        assert_eq!(utf8_byte_to_utf16_offset(source, hiragana_start), 1);
-        assert_eq!(utf8_byte_to_utf16_offset(source, emoji_start), 2);
-        assert_eq!(utf8_byte_to_utf16_offset(source, latin_b_start), 4);
-        assert_eq!(utf8_byte_to_utf16_offset(source, source.len() as u32), 5);
-    }
-}
-
-/// Parse CSS options from JsValue
-pub(crate) fn parse_css_options(options: JsValue) -> CssCompileOptions {
-    let scope_id = js_sys::Reflect::get(&options, &JsValue::from_str("scopeId"))
-        .ok()
-        .and_then(|v| v.as_string())
-        .map(Into::into);
-
-    let scoped = js_sys::Reflect::get(&options, &JsValue::from_str("scoped"))
-        .ok()
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    let minify = js_sys::Reflect::get(&options, &JsValue::from_str("minify"))
-        .ok()
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    let source_map = js_sys::Reflect::get(&options, &JsValue::from_str("sourceMap"))
-        .ok()
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    let css_modules = js_sys::Reflect::get(&options, &JsValue::from_str("cssModules"))
-        .ok()
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    let custom_media = js_sys::Reflect::get(&options, &JsValue::from_str("customMedia"))
-        .ok()
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    let filename = js_sys::Reflect::get(&options, &JsValue::from_str("filename"))
-        .ok()
-        .and_then(|v| v.as_string())
-        .map(Into::into);
-
-    // Parse targets
-    let targets = js_sys::Reflect::get(&options, &JsValue::from_str("targets"))
-        .ok()
-        .and_then(|v| {
-            if v.is_undefined() || v.is_null() {
-                return None;
-            }
-            Some(CssTargets {
-                chrome: js_sys::Reflect::get(&v, &JsValue::from_str("chrome"))
-                    .ok()
-                    .and_then(|v| v.as_f64())
-                    .map(|v| v as u32),
-                firefox: js_sys::Reflect::get(&v, &JsValue::from_str("firefox"))
-                    .ok()
-                    .and_then(|v| v.as_f64())
-                    .map(|v| v as u32),
-                safari: js_sys::Reflect::get(&v, &JsValue::from_str("safari"))
-                    .ok()
-                    .and_then(|v| v.as_f64())
-                    .map(|v| v as u32),
-                edge: js_sys::Reflect::get(&v, &JsValue::from_str("edge"))
-                    .ok()
-                    .and_then(|v| v.as_f64())
-                    .map(|v| v as u32),
-                ios: js_sys::Reflect::get(&v, &JsValue::from_str("ios"))
-                    .ok()
-                    .and_then(|v| v.as_f64())
-                    .map(|v| v as u32),
-                android: js_sys::Reflect::get(&v, &JsValue::from_str("android"))
-                    .ok()
-                    .and_then(|v| v.as_f64())
-                    .map(|v| v as u32),
-            })
-        });
-
-    CssCompileOptions {
-        scope_id,
-        scoped,
-        minify,
-        source_map,
-        targets,
-        filename,
-        custom_media,
-        css_modules,
-    }
-}
-
-/// SFC compile result for WASM
-#[derive(Serialize)]
-pub struct SfcWasmResult {
-    pub descriptor: SfcDescriptorWasm,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub template: Option<CompileResult>,
-    pub script: SfcScriptResult,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub css: Option<String>,
-    pub errors: Vec<String>,
-    pub warnings: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "bindingMetadata")]
-    pub binding_metadata: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Vec::is_empty", rename = "macroArtifacts")]
-    pub macro_artifacts: Vec<SfcMacroArtifactWasm>,
-}
-
-/// Script compilation result
-#[derive(Serialize)]
-pub struct SfcScriptResult {
-    pub code: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub bindings: Option<serde_json::Value>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcMacroArtifactWasm {
-    pub kind: String,
-    pub name: String,
-    pub source: String,
-    pub content: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub module_code: Option<String>,
-    pub start: usize,
-    pub end: usize,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcDescriptorWasm {
-    pub filename: String,
-    pub source: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub template: Option<SfcTemplateBlockWasm>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub script: Option<SfcScriptBlockWasm>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub script_setup: Option<SfcScriptBlockWasm>,
-    pub styles: Vec<SfcStyleBlockWasm>,
-    pub custom_blocks: Vec<SfcCustomBlockWasm>,
-    pub css_vars: Vec<String>,
-    pub slotted: bool,
-    pub should_force_reload: bool,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcTemplateBlockWasm {
-    pub content: String,
-    pub loc: SfcBlockLocationWasm,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lang: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub src: Option<String>,
-    pub attrs: BTreeMap<String, String>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcScriptBlockWasm {
-    pub content: String,
-    pub loc: SfcBlockLocationWasm,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lang: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub src: Option<String>,
-    pub setup: bool,
-    pub attrs: BTreeMap<String, String>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcStyleBlockWasm {
-    pub content: String,
-    pub loc: SfcBlockLocationWasm,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lang: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub src: Option<String>,
-    pub scoped: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub module: Option<String>,
-    pub attrs: BTreeMap<String, String>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcCustomBlockWasm {
-    pub r#type: String,
-    pub content: String,
-    pub attrs: BTreeMap<String, String>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SfcBlockLocationWasm {
-    pub start: usize,
-    pub end: usize,
-    pub tag_start: usize,
-    pub tag_end: usize,
-    pub start_line: usize,
-    pub start_column: usize,
-    pub end_line: usize,
-    pub end_column: usize,
-}
-
-fn attrs_to_wasm(
-    attrs: &vize_carton::FxHashMap<std::borrow::Cow<'_, str>, std::borrow::Cow<'_, str>>,
-) -> BTreeMap<String, String> {
-    attrs
-        .iter()
-        .map(|(key, value)| (key.to_string(), value.to_string()))
-        .collect()
-}
-
-fn loc_to_wasm(loc: &vize_atelier_sfc::BlockLocation) -> SfcBlockLocationWasm {
-    SfcBlockLocationWasm {
-        start: loc.start,
-        end: loc.end,
-        tag_start: loc.tag_start,
-        tag_end: loc.tag_end,
-        start_line: loc.start_line,
-        start_column: loc.start_column,
-        end_line: loc.end_line,
-        end_column: loc.end_column,
-    }
-}
-
-fn template_block_to_wasm(block: &vize_atelier_sfc::SfcTemplateBlock<'_>) -> SfcTemplateBlockWasm {
-    SfcTemplateBlockWasm {
-        content: block.content.to_string(),
-        loc: loc_to_wasm(&block.loc),
-        lang: block.lang.as_ref().map(|value| value.to_string()),
-        src: block.src.as_ref().map(|value| value.to_string()),
-        attrs: attrs_to_wasm(&block.attrs),
-    }
-}
-
-fn script_block_to_wasm(block: &vize_atelier_sfc::SfcScriptBlock<'_>) -> SfcScriptBlockWasm {
-    SfcScriptBlockWasm {
-        content: block.content.to_string(),
-        loc: loc_to_wasm(&block.loc),
-        lang: block.lang.as_ref().map(|value| value.to_string()),
-        src: block.src.as_ref().map(|value| value.to_string()),
-        setup: block.setup,
-        attrs: attrs_to_wasm(&block.attrs),
-    }
-}
-
-fn style_block_to_wasm(block: &vize_atelier_sfc::SfcStyleBlock<'_>) -> SfcStyleBlockWasm {
-    SfcStyleBlockWasm {
-        content: block.content.to_string(),
-        loc: loc_to_wasm(&block.loc),
-        lang: block.lang.as_ref().map(|value| value.to_string()),
-        src: block.src.as_ref().map(|value| value.to_string()),
-        scoped: block.scoped,
-        module: block.module.as_ref().map(|value| value.to_string()),
-        attrs: attrs_to_wasm(&block.attrs),
-    }
-}
-
-fn custom_block_to_wasm(block: &vize_atelier_sfc::SfcCustomBlock<'_>) -> SfcCustomBlockWasm {
-    SfcCustomBlockWasm {
-        r#type: block.block_type.to_string(),
-        content: block.content.to_string(),
-        attrs: attrs_to_wasm(&block.attrs),
-    }
-}
-
-fn descriptor_to_wasm(descriptor: &SfcDescriptor<'_>) -> SfcDescriptorWasm {
-    SfcDescriptorWasm {
-        filename: descriptor.filename.to_string(),
-        source: descriptor.source.to_string(),
-        template: descriptor.template.as_ref().map(template_block_to_wasm),
-        script: descriptor.script.as_ref().map(script_block_to_wasm),
-        script_setup: descriptor.script_setup.as_ref().map(script_block_to_wasm),
-        styles: descriptor.styles.iter().map(style_block_to_wasm).collect(),
-        custom_blocks: descriptor
-            .custom_blocks
-            .iter()
-            .map(custom_block_to_wasm)
-            .collect(),
-        css_vars: descriptor
-            .css_vars
-            .iter()
-            .map(|value| value.to_string())
-            .collect(),
-        slotted: descriptor.slotted,
-        should_force_reload: descriptor.should_force_reload,
-    }
-}
-
-fn macro_artifact_to_wasm(artifact: &SfcMacroArtifact) -> SfcMacroArtifactWasm {
-    SfcMacroArtifactWasm {
-        kind: artifact.kind.to_string(),
-        name: artifact.name.to_string(),
-        source: artifact.source.to_string(),
-        content: artifact.content.to_string(),
-        module_code: artifact.module_code.as_ref().map(ToString::to_string),
-        start: artifact.start,
-        end: artifact.end,
-    }
-}
+use super::options::{parse_compiler_options, parse_css_options};
+use super::serde::{to_js_value, to_json_js_value};
+use super::sfc_types::{
+    SfcScriptResult, SfcWasmResult, descriptor_to_wasm, macro_artifact_to_wasm,
+};
 
 /// WASM Compiler instance
 #[wasm_bindgen]
